+++ CHANGELOG.md
@@ -31,12 +31,13 @@
     <summary>Changes</summary>
 
 - New Features
-  - Added /admin/logout and basic functionality
-  - Added /admin/login and basic functionality
+  - Added POST /admin/logout and basic functionality
+  - Added POST /admin/login and basic functionality
   - Added Account schema layout for account creation
   - Added Account controller for route logic
   - Added authorization header as current form of auth
   - Added validation for login
+  - Added JWT authorization middleware for protected routes
 - Changes
   - Routes now correctly use /api as a prefix
   - Changed routing system to use express.Router()